moving the code of the routes to the celebrities.js file
