+++ starter-code/app.js
@@ -52,7 +52,10 @@ app.locals.title = 'Express - Generated with IronGenerator';
 
 
 const index = require('./routes/index');
+const celebrities = require('./routes/celebrities');
+
 app.use('/', index);
+app.use('/celebrities', celebrities);
 
 
 module.exports = app;
