+++ starter-code/views/index.hbs
@@ -1,2 +1,2 @@
-<h1>{{title}}</h1>
-<p>Welcome to {{title}}</p>
+<h1>Movies & Celebrities</h1>
+<a href="/celebrities">Celebrities</a>
